Moved level initialzation to a new object
This allows us to just work with loyal as a controling object and not a
code hub.

Currently the code works but the future is to move all instances of
the Static levels to the Level intializer and have them get set in the
decision objects run by each level.