Changed the battleController MVC framework
The code should now work with a single reference to BattleController
which Loyal will call tick on and it will check Battle for information
feeding it indexes which allow Battle to select the proper abilities.

Also fixed the MapPlayer testing code which initialized the Decision for
Overworld switching from Village to Overworld.